% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_button.R
\name{dropdown_button}
\alias{dropdown_button}
\title{Dropdown button}
\usage{
dropdown_button(
  ...,
  id,
  label = NULL,
  class = NULL,
  split = FALSE,
  outline = FALSE,
  dark = FALSE,
  size = c("m", "s", "l"),
  type = c("primary", "secondary", "success", "info", "warning", "danger"),
  direction = c("down", "up", "left", "right")
)
}
\description{
Dropdown button
}
