% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_button.R
\name{dropdown_divider}
\alias{dropdown_divider}
\title{Create a Bootstrap dropdown menu divider}
\usage{
dropdown_divider()
}
\value{
An HTML list item tag for use inside [dropdown_button()].
}
\description{
Creates a horizontal rule to visually separate groups of items inside a
[dropdown_button()] menu.
}
\examples{
# No run
}
\seealso{
[dropdown_button()], [dropdown_item()], [dropdown_header()]
}
