% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_button.R
\name{dropdown_header}
\alias{dropdown_header}
\title{Create a Bootstrap dropdown menu header}
\usage{
dropdown_header(label)
}
\arguments{
\item{label}{The text to display as the group header.}
}
\value{
An HTML list item tag for use inside [dropdown_button()].
}
\description{
Creates a non-interactive header label to group related items inside a
[dropdown_button()] menu.
}
\examples{
# No run
}
\seealso{
[dropdown_button()], [dropdown_item()], [dropdown_divider()]
}
