% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_input.R
\name{email_input}
\alias{email_input}
\title{Create a email input control}
\usage{
email_input(
  id,
  label = NULL,
  value = NULL,
  floating = FALSE,
  placeholder = NULL,
  size = c("m", "s", "l"),
  disabled = FALSE,
  class = NULL,
  readonly = FALSE,
  update_on = c("change", "blur")
)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{floating}{If true, the label is floated over the input field.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{size}{Controls the size of the input field.}

\item{disabled}{If true, the input field is shaded out and cannot be modified.}

\item{class}{Additional Bootstrap utilities to manipulate the input control.}

\item{readonly}{If true, the input field cannot be modified but appears active.}

\item{update_on}{something....}
}
\value{
An email input control that can be added to a UI definition.
}
\description{
Create an input control for entry of unstructured text values for email
}
\examples{
# No run
}
