% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal_body}
\alias{modal_body}
\title{Create a Bootstrap modal body}
\usage{
modal_body(..., id = NULL, class = NULL)
}
\arguments{
\item{...}{Content elements to include in the modal body.}

\item{id}{An optional HTML id for the body element.}

\item{class}{Additional CSS classes to apply to the body element.}
}
\value{
An HTML tag that can be passed to the `body` argument of
  [modal_dialog()].
}
\description{
Creates the body section of a Bootstrap modal to hold the main content.
}
\examples{
# No run
}
\seealso{
[modal_dialog()]
}
