% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal_button}
\alias{modal_button}
\title{Create a Bootstrap modal dismiss button}
\usage{
modal_button(id = NULL, label, class = NULL)
}
\arguments{
\item{id}{An optional HTML id for the button element.}

\item{label}{The text label for the button.}

\item{class}{Additional CSS classes to apply to the button element.}
}
\value{
An HTML tag that can be passed to [modal_footer()].
}
\description{
Creates a button that dismisses the modal when clicked. Intended for use
inside [modal_footer()].
}
\examples{
# No run
}
\seealso{
[modal_footer()], [modal_dialog()]
}
