% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal_footer}
\alias{modal_footer}
\title{Create a Bootstrap modal footer}
\usage{
modal_footer(..., id = NULL, class = NULL)
}
\arguments{
\item{...}{Content elements to include in the modal footer, typically buttons.}

\item{id}{An optional HTML id for the footer element.}

\item{class}{Additional CSS classes to apply to the footer element.}
}
\value{
An HTML tag that can be passed to the `footer` argument of
  [modal_dialog()].
}
\description{
Creates the footer section of a Bootstrap modal, typically used to hold
action buttons such as [modal_button()] or [toggle_button()].
}
\examples{
# No run
}
\seealso{
[modal_dialog()], [modal_button()]
}
