% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvas_body}
\alias{offcanvas_body}
\title{Create a Bootstrap offcanvas body}
\usage{
offcanvas_body(..., id = NULL, class = NULL)
}
\arguments{
\item{...}{Content elements to include in the offcanvas body.}

\item{id}{An optional HTML id for the body element.}

\item{class}{Additional CSS classes to apply to the body element.}
}
\value{
An HTML tag that can be passed to the `body` argument of
  [offcanvas_dialog()].
}
\description{
Creates the body section of a Bootstrap offcanvas panel to hold the main
content.
}
\examples{
# No run
}
\seealso{
[offcanvas_dialog()]
}
