% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvas_button}
\alias{offcanvas_button}
\title{Create a Bootstrap offcanvas toggle button}
\usage{
offcanvas_button(
  target_id,
  id = NULL,
  label = NULL,
  class = NULL,
  toggle_type = c("offcanvas", "collapse")
)
}
\arguments{
\item{target_id}{The HTML id of the [offcanvas_dialog()] to toggle.}

\item{id}{An optional HTML id for the button element.}

\item{label}{The text label for the button.}

\item{class}{Additional CSS classes to apply to the button element.}

\item{toggle_type}{Currently unused. Reserved for future use.}
}
\value{
An HTML button tag that can be added to a UI definition.
}
\description{
Creates a button that opens an offcanvas panel when clicked. For a more
flexible alternative that also supports collapse and modal toggling, see
[toggle_button()].
}
\examples{
# No run
}
\seealso{
[offcanvas_dialog()], [toggle_button()]
}
