% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvas_header}
\alias{offcanvas_header}
\title{Create a Bootstrap offcanvas header}
\usage{
offcanvas_header(..., id = NULL, title = NULL)
}
\arguments{
\item{...}{Currently unused. Reserved for future use.}

\item{id}{An optional HTML id for the title element.}

\item{title}{The text to display as the offcanvas panel title.}
}
\value{
An HTML tag that can be passed to the `header` argument of
  [offcanvas_dialog()].
}
\description{
Creates the header section of a Bootstrap offcanvas panel, containing a
title and a close button.
}
\examples{
# No run
}
\seealso{
[offcanvas_dialog()]
}
