% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_input.R
\name{select_input}
\alias{select_input}
\title{Create a select list input control}
\usage{
select_input(
  id,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE,
  width = NULL,
  size = NULL,
  disable = FALSE
)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control.}

\item{choices}{List of values to select from. If elements of the list are
named, those names are used as the displayed labels.}

\item{selected}{The initially selected value. If `NULL`, defaults to the
first element of `choices` (or no selection for `multiple = TRUE`).}

\item{multiple}{If `TRUE`, multiple items can be selected at once.}

\item{selectize}{If `TRUE`, the select widget is enhanced using the
selectize.js library.}

\item{width}{The width of the input, e.g. `"200px"` or `"50%"`.}

\item{size}{Number of items to show in the list box when `selectize = FALSE`.}

\item{disable}{If `TRUE`, the input is shaded out and cannot be modified.}
}
\value{
A select list control that can be added to a UI definition.
}
\description{
A wrapper around [shiny::selectInput()] that creates a select list to choose
a single or multiple items from a list of values, with an added `disable`
option not available in the base function.
}
\examples{
# No run
}
\seealso{
[shiny::selectInput()]
}
